fix bug
fix bug

+++ Loon-0.5-Alpha/Loon-Neo/src/loon/geom/Dimension.java
@@ -72,7 +72,7 @@ public class Dimension {
 		return dirty;
 	}
 
-	public boolean contains(int x, int y) {
+	public boolean contains(float x, float y) {
 		return x >= 0 && x < width && y >= 0 && y < height;
 	}
